backend/requirements.txt
# Reality Simulator Engine — Backend Dependencies

# Phase 2+ (economy logging, optional):
# tabulate>=0.9.0

# Phase 4+ (Groq AI integration):
groq>=0.4.2
python-dotenv>=1.0.0

# Phase 6+ (REST API for frontend):
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Optional (data persistence):
# sqlitedict>=2.1.0
